// ui/src/apis/proto/openpay/v1/ledger.ts
export enum PostingDirection {
  POSTING_DIRECTION_UNSPECIFIED = "POSTING_DIRECTION_UNSPECIFIED",
  POSTING_DIRECTION_DEBIT = "POSTING_DIRECTION_DEBIT",
  POSTING_DIRECTION_CREDIT = "POSTING_DIRECTION_CREDIT",
  UNRECOGNIZED = "UNRECOGNIZED",
}